[Ch.9] Extract the NavHost into RallyNavHost
